setup retrofit added repositories to Hotels, Rooms and Booking added usecases to Hotels, Rooms and Booking added viewmodels to Hotels, Rooms and Booking showed data from api on screen

+++ app/src/main/java/com/example/hotelapplication/mappers/DefaultBookingDataMapper.kt
@@ -1,9 +1,9 @@
 package com.example.hotelapplication.mappers
 
-import com.example.booking.domain.dao.BookingDataDao
-import com.example.booking.domain.mappers.BookingDataMapper
+import com.example.core.dao.booking.BookingDataDao
+import com.example.core.mappers.BookingDataMapper
 import com.example.core.Constants
-import com.example.data.network.dto.booking_info.BookingDataDto
+import com.example.core.entities.booking_info.BookingDataDto
 import javax.inject.Inject
 
 class DefaultBookingDataMapper @Inject constructor() : BookingDataMapper {
@@ -11,17 +11,17 @@ class DefaultBookingDataMapper @Inject constructor() : BookingDataMapper {
         id = bookingDataDto.id ?: -1L,
         hotelName = bookingDataDto.hotel_name ?: Constants.ERROR_STRING,
         hotelAddress = bookingDataDto.hotel_adress ?: Constants.ERROR_STRING,
-        rating = (bookingDataDto.horating ?: -1) as Int,
+        rating = bookingDataDto.horating ?: -1,
         ratingName = bookingDataDto.rating_name ?: Constants.ERROR_STRING,
         departure = bookingDataDto.departure ?: Constants.ERROR_STRING,
         arrivalCountry = bookingDataDto.arrival_country ?: Constants.ERROR_STRING,
         tourDateStart = bookingDataDto.tour_date_start ?: Constants.ERROR_STRING,
         tourDateStop = bookingDataDto.tour_date_stop ?: Constants.ERROR_STRING,
-        numberOfNights = (bookingDataDto.number_of_nights ?: -1) as Int,
+        numberOfNights = bookingDataDto.number_of_nights ?: -1,
         room = bookingDataDto.room ?: Constants.ERROR_STRING,
         nutrition = bookingDataDto.nutrition ?: Constants.ERROR_STRING,
-        tourPrice = (bookingDataDto.tour_price ?: -1F) as Float,
-        fuelCharge = (bookingDataDto.fuel_charge ?: -1F) as Float,
-        serviceCharge = (bookingDataDto.service_charge ?: -1F) as Float
+        tourPrice = bookingDataDto.tour_price ?: -1F,
+        fuelCharge = bookingDataDto.fuel_charge ?: -1F,
+        serviceCharge = bookingDataDto.service_charge ?: -1F
     )
 }
