+++ app/src/main/java/com/example/hotelapplication/mappers/DefaultHotelMapper.kt
@@ -1,20 +1,20 @@
 package com.example.hotelapplication.mappers
 
 import com.example.core.Constants
-import com.example.data.network.dto.hotel.HotelDto
-import com.example.hotel.domain.dao.AboutTheHotelDao
-import com.example.hotel.domain.dao.HotelDao
-import com.example.hotel.domain.mappers.HotelMapper
+import com.example.core.entities.hotel.HotelDto
+import com.example.core.dao.hotel.AboutTheHotelDao
+import com.example.core.dao.hotel.HotelDao
+import com.example.core.mappers.HotelMapper
 import javax.inject.Inject
 
 class DefaultHotelMapper @Inject constructor() : HotelMapper {
     override fun mapDtoToDao(hotelDto: HotelDto) = HotelDao(
-        id = (hotelDto.id ?: -1F) as Long,
+        id = hotelDto.id ?: -1L,
         name = hotelDto.name ?: Constants.ERROR_STRING,
         address = hotelDto.adress ?: Constants.ERROR_STRING,
-        minimalPrice = (hotelDto.minimal_price ?: -1F) as Float,
+        minimalPrice = hotelDto.minimal_price ?: -1F,
         priceForIt = hotelDto.price_for_it ?: Constants.ERROR_STRING,
-        rating = (hotelDto.rating ?: -1) as Int,
+        rating = hotelDto.rating ?: -1,
         ratingName = hotelDto.rating_name ?: Constants.ERROR_STRING,
         images = hotelDto.image_urls,
         aboutTheHotel = AboutTheHotelDao(
